Update parameters for modern systems
I've changed the adjustable defines to better reflect the performance
characteristics of modern computers. The performance gap between memory
and hard disk is far wider than before, so I'm changing RUNTM and STOPTM
to stop shortly after the benchmark causes the system to thrash on the
disk. This will also avoid excessive load on SSDs when the benchmark has
used up all main memory in the computer.

NSAMP has been adjusted to allow more samples. This increases the size
of a statically allocated array in the code, which should not be an
issue on modern computers.

RUNTM is reduced to 0.5 to speed early stages of the benchmark. System
timers are typically at least 100 HZ today and often run at 1000 HZ, so
this should not be an issue.

I'm also noting the changes I've made at the top of the file.